refactor: Move plugins to their own directories
