+++ packages/wrangler/src/deployment-bundle/esbuild-plugins/cloudflare-internal.ts
@@ -1,0 +1,13 @@
+import type { Plugin } from "esbuild";
+
+/**
+ * An esbuild plugin that will mark any `cloudflare:...` imports as external.
+ */
+export const cloudflareInternalPlugin: Plugin = {
+	name: "cloudflare javascript Plugin",
+	setup(pluginBuild) {
+		pluginBuild.onResolve({ filter: /^cloudflare:.*/ }, () => {
+			return { external: true };
+		});
+	},
+};
